// src/libs/models/FixedSwapContract.d.ts
export interface FixedSwapContractPurchase {
    _id: string,
    amount: string,
    purchaser: string,
    ethAmount: string,
    timestamp: Date,
    wasFinalized: boolean,
    reverted: boolean,
    validAfterDate?: Date,
    refunded: boolean,
    refundStart: Date,
	refundEnd: Date
}